Merge branch 'master' of github.com:zololabs/zolodeck

# Please enter a commit message to explain why this merge is necessary,
# especially if it merges an updated upstream into a topic branch.
#
# Lines starting with '#' will be ignored, and an empty message aborts
# the commit.
# Please enter a commit message to explain why this merge is necessary,
# especially if it merges an updated upstream into a topic branch.
#
# Lines starting with '#' will be ignored, and an empty message aborts
# the commit.

+++ Rakefile
@@ -38,7 +38,9 @@ namespace :uberjar do
 
     sh "lein clean"
     if !debug
-      sh "rm devops/config/prod/zolo.clj"
+      if File.exists? "devops/config/prod/zolo.clj"
+        sh "rm devops/config/prod/zolo.clj"
+      end
       Rake::Task["prod:config:download"].invoke
     end
     sh "lein with-profile storm uberjar zolo.storm.facebook"
